qdrant/requirements-ingest.txt
# Dependencies for qdrant/ingest_news.py (install on top of project venv):
#
#   pip install -r qdrant/requirements-ingest.txt
#
# Default embeddings: local jinaai/jina-embeddings-v3 (sentence-transformers; first run downloads weights).
#
qdrant-client>=1.12.0
httpx>=0.27.0
tqdm>=4.67.0
python-dotenv>=1.0.0
sentence-transformers>=3.1.0

# Optional: NEWS_EMBED_BACKEND=openai uses EMBEDDING_BASE_URL + EMBEDDING_API_KEY (no extra package beyond httpx).
